requirements.txt
psycopg==3.2.3
requests==2.32.3
aiohttp==3.10.11
pytest-asyncio==0.24.0
aioresponses
boto3==1.35.69